requirements.txt
absl-py==2.1.0
brotlipy==0.7.0
cachetools==5.3.3
certifi @ file:///croot/certifi_1707229174982/work/certifi
cffi @ file:///tmp/build/80754af9/cffi_1636541934635/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1661170624537/work
contourpy==1.1.1
cryptography @ file:///croot/cryptography_1677533068310/work
cycler==0.12.1
filelock @ file:///croot/filelock_1700591183607/work
fonttools==4.50.0
fsspec @ file:///croot/fsspec_1701286474621/work
future @ file:///croot/future_1677599870788/work
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645455532332/work
google-auth==2.29.0
google-auth-oauthlib==1.0.0
grpcio==1.62.1
idna @ file:///croot/idna_1666125576474/work
imageio==2.34.0
imgaug==0.4.0
importlib_metadata==7.1.0
importlib_resources==6.4.0
Jinja2 @ file:///croot/jinja2_1706733616596/work
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.3
Markdown==3.6
MarkupSafe @ file:///croot/markupsafe_1704205993651/work
matplotlib==3.7.5
mpmath @ file:///croot/mpmath_1690848262763/work
networkx @ file:///croot/networkx_1690561992265/work
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1675642515540/work
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opencv-python==4.9.0.80
packaging==24.0
pandas==2.0.3
Pillow==9.4.0
progressbar2==4.2.0
protobuf==5.26.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyOpenSSL @ file:///croot/pyopenssl_1677607685877/work
pyparsing==3.1.2
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil==2.9.0.post0
python-utils==3.5.2
pytz==2024.1
PyWavelets==1.4.1
PyYAML @ file:///croot/pyyaml_1670514731622/work
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1673863902341/work
requests-oauthlib==2.0.0
rsa==4.9
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
shapely==2.0.3
six==1.16.0
sympy @ file:///croot/sympy_1701397643339/work
tensorboard==2.14.0
tensorboard-data-server==0.7.2
threadpoolctl==3.4.0
tifffile==2023.7.10
torch==2.2.1
torchaudio==2.2.1
torchvision==0.17.1
tqdm==4.66.2
triton==2.2.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1678635778344/work
Werkzeug==3.0.1
zipp==3.18.1
